Added show function and as.data.frame for samples and chronologies to display clearer when navigating through site object.
Changed to cli library for hashing, other libraries are not hashing properly the complete R object.

added as.character() to samples and specimens dataframe as some download objects are coming back empty, triggering the creation of a null object with an id (API also shows null objects - not sure how to filter this out)

+++ R/01_classDefinitions.R
@@ -414,7 +414,7 @@ setClass(
             depositionalenvironment = "character",
             datasets = "datasets",
             chronologies = "chronologies",
-            defaultchronology = "integer"),
+            defaultchronology = "numeric"),
   prototype = list(collectionunitid = NA_integer_,
                    notes = NA_character_,
                    handle = NA_character_,
